+++ com.unity.mobile.android-logcat/Editor/UI/AndroidLogcatRegexList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0af400dccbb10d041a0ba66c8e9aa6c2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
