+++ com.unity.mobile.android-logcat/Editor/UI/AndroidLogcatSymbolList.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 68057dd8ee115884fbb531c56654c5af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
